fix: Rename map.py mean_ap.py to avoid collision with builtin map (#908)
